Work out what a caller may download, from the PWRD entry they match
Groundwork for the transfer limits, which have never been enforced: nothing
read a security level beyond its time limit, and the PWRD import quietly threw
away the base baud rate and the batch limit.

The arithmetic follows checkratio/checklimit and checkdlfiles in the original
source rather than the manuals, which leave most of this undocumented. A ratio
is held in tenths, so the fields say so now. Zero disables a ratio or a limit,
32767 K a day means unlimited, and a caller who has never uploaded counts as
having uploaded one so the ratio still yields its share instead of dividing by
zero. Credits stand in for uploads that have not happened yet.

The daily download figure is signed, as PCBoard's DailyDnldBytes is: an upload
earns credit against the day's allowance and pushes it below zero. That also
settles a bug in ADJBYTES, where subtracting went through `bytes as u64` on a
negative number and saturated the counter to zero instead of lowering it.

Nothing enforces any of this yet.

+++ crates/icboard/src/tui.rs
@@ -357,7 +357,7 @@ impl Tui {
                         status_bar_info.msg_read,
                         status_bar_info.dn,
                         status_bar_info.dnbytes,
-                        status_bar_info.today_dn as i64 + status_bar_info.today_ul as i64,
+                        status_bar_info.today_dn + status_bar_info.today_ul as i64,
                     ))
                     .style(Style::new().fg(DOS_BLACK).bg(DOS_LIGHT_GRAY)),
                 ]);
@@ -465,7 +465,7 @@ pub struct StatusBarInfo {
     pub msg_left: u64,
     pub msg_read: u64,
     pub today_ul: u64,
-    pub today_dn: u64,
+    pub today_dn: i64,
     pub bus_phone: String,
     pub home_phone: String,
     pub email: String,
